rust: Add "reborrow" API, use for live path (+1 cancellable bit)
I was going to do some further oxidation and as I was typing out the
whole `Pin<&mut T>` thing again I was thinking about how in other
places we were passing plain Rust references i.e. `&Foo` not
`Pin<&mut Foo>`.

With this, it works fine to pass borrowed glib-rs types across via
cxx-rs.  I am not totally sure how I got it into my head originally
that we needed `Pin<&mut T>` here.

In particular, this case of C/C++ → Rust is basically the same as
the use case of GObject signals (e.g. Rust GTK+ application catching
a resize signal).  We can use
https://docs.rs/glib/latest/glib/translate/trait.FromGlibPtrBorrow.html
which is designed for this.

Another way to look at this is: We don't need to bump the gobject
reference count here - we have the Rust lifetime rules to protect us.

The `Pin<&mut T>` bits for cxx-rs are designed to map to
C++ mutable vs `const`, but that doesn't apply at all to plain C where
`const` is even more useless than in C++, and hence the glib-rs Rust
code just passes everything as shared `&` references and not `&mut`.

After this lands I think we can slowly convert the rest of our FFI
to use this, and drop most of the `Pin<&mut T>` bits.

I started out testing this by changing one random `GCancellable` bit
in composepost, but then decided to try a full port of the `live.rs`
API.