diff: Drop use of Pin<&mut T> for ostree repo
This one is subtle.  Again here, cxx.rs wants to try to match up
rust `&` and `&mut` with C++ `const &` vs `&`.

However...on the C side, we take raw `OstreeRepo *` pointers that
all implicitly support "interior mutability".

We can then (safely AIUI) use `const_cast` to cast away the `const`
that cxx wants to get a reference, which we then turn into a raw
C pointer which is passed to the C side.