app/override: Always treat NEVRA as frozen
If we're given a NEVRA and `--from`, handle it right away as a frozen
local override instead of converting it to a remote override (which
we'll add client-side support for soon).

A NEVRA pin is functionally the same as a local override after the repo
fetching, and it's silly to always try to refetch the same NEVRA from
the remote every upgrade.