tree: Rename local replacement-related variables
Rename all the variables associated with the commit metadata key
`rpmostree.replaced-base-packages` to specify `local` to be more correct
as prep for introducing remote replacements. It'd be nice to rename the
commit metadata key name as well, but that'd require a clientlayer
version bump and extra handling in the bits that read it which doesn't
seem worth it.