core: Remove redundant check from `check_goal_solution`
This check is in fact a tautology because the way `pkgs_to_replace` is
constructed, every package in it must have been in `replaced_pkgnames`.

I think the original intent here was to actually check the inverse: that
every package in `replaced_pkgnames` is also in `pkgs_to_replace`.
However, I don't think this is worth checking because `replaced_pkgames`
itself is derived from packages we marked for installation. If there's a
mismatch, that would be a libdnf (or more likely libsolv) bug.