core: Strengthen replacement handling in `check_goal_solution`
We were a bit wishy-washy in `check_goal_solution` when looking at
replaced packages where we tried to handle override replacements that
obsolete packages of a different name.

I don't believe this can happen in practice because our overrides are
purely package name based. If indeed a base package override also
obsoletes another base package of a different name, then that base
package would also need to be listed as an override removal.

So be more thorough here when looking for the replaced package so that
we don't mistakenly pick the incorrect one.