app/status: Simplify local replacement printing
Now that we know that overrides always have matching package names, we
can simplify the code here by deleting the `g_str_equal` if-condition.
In fact, the `else` branch of that wasn't really useful because we
regressed at some point in printing the contents of `buf`.

While we're here, drop unused variables, simplify variable names and
handling of `buf`.