compose: Use `GVariantDict`, not manual hash table
There's a dedicated type for a mapping from "string to variant" in
`GVariantDict`.  Use it instead of rolling our own with `GHashTable`.
The advantage here is improved ergonomics from the C side; e.g.
we can directly pass C strings for values and have the variant
allocated internally.

But more importantly, this is prep for oxidizing things here because
we can pass `GVariantDict` to/from Rust easily, not so with
`GHashTable`.