app/status: Reuse `print_values` to print rpmmd repos
That way, the knowledge of handling `TREE_RIGHT` and `TREE_BRANCH` to
print a list lives in just one place.

While we're here, just merge `print_commit` and `print_origin_repos`.
The only special thing here is that `print_values` doesn't print
anything if no values exist, so we print the commit checksum ourselves
if no repos need to be printed.