rustix: use libc backend
See rationale in https://github.com/bytecodealliance/rustix/issues/215
I wrote it so that we can use it here.

And specifically this will fix a currently s390x-specific build failure
because `rustix` only depends on the `errno` crate if it's using the
libc backend, and this creates currently `s390x` specific build failures
for us.