composepost: Add and use helpers for reading/writing absolute symlinks
See https://pagure.io/releng/failed-composes/issue/3660

The port to cap-std regressed only the *non-unified* core case
for tmpfiles.d translation.  Here cap-std's `read_link()` API correctly
checks for relative paths for the link name, but it *also* bombs
out if the link *target* is absolute, even if we're not going to
follow it.

I think this is arguably buggy behavior.  But, that's a debate
for a future time.

Now, this bug was masked in our CI because it doesn't occur in
unified core mode where we handle tmpfiles.d translation as part
of package imports.

Add and use internal helpers for reading (and writing, for tests)
absolute symlinks.