app/kargs: Use common function to run transaction
Make `rpm-ostree kargs` like the others and use
`rpmostree_transaction_client_run` to complete the transaction.

This fixes a bug at the same time: if `--reboot` is used, we don't want
to e.g. query the daemon or we'll risk racing with it shutting down.
This in turn can cause unsightly assertion errors as the machine goes
down.

It also fixes minor UX issues like not printing the reboot hint when we
are in fact rebooting.

Reported-by: Dusty Mabe <dusty@dustymabe.com>