docs/HACKING: Rewrite dev-environment setup
Adds a separate section to explain how to set up a development
environment. It explains what needs to be done when working either
natively, inside a toolbx container or when using a "cattle" container
(fcos-buildroot). Special emphasis is put on the ci scripts that install
all the dependencies for the user.

The contents strictly related to building and testing are now beneath a
new heading "Building and testing".