rust: Update to memfd 0.6.1
Notably this pulls in the latest rustix, which means we only
have one copy linked in process.  rustix is a quite nontrivial
project (roughly equivalent in scope to a C library).

In the future I plan to use e.g. `cargo deny` to ensure we don't
get in this situation again.