Give a better error for `ex module` in a container
The modularity code is native Rust, which is great, but means it
currently bypasses the infrastructure we added mostly on the C++
side (where most CLI entrypoint logic is) to detect the case of
ostree host/DBus versus container etc.

The Rust side already has a nice `ClientConnection` object that
basically expresses the need for an ostree host.  And we already
had a Rust API that's useful to error out here; combine those
two and tweak the error message.

(A future patch will change `ex module` to actually work, but
 this is a correct starting point)