treefile: Retain directory path, not dfd
This is part of https://github.com/coreos/rpm-ostree/issues/3832
but also just a general code cleanup - in general, using file descriptors
is great, but libdnf doesn't support them, so we are awkwardly holding
onto a dfd and using `/proc` to make it a path.

Since our treefile input is today always a filesystem path, we might
as well expose that back - we're just holding onto the parent directory
*path*.

If in the future we converted the treefile code to accept a cap-std
`Dir` as input, then in theory we'd likely want to push for libdnf
to have `dnf_context_set_repo_dfd` or so.

But, this simplifies the code today.