Drop use of `curl`, use `reqwest` instead
A while ago when I added the `xmlrpc` dependency in order to talk
to koji, I didn't notice that it depended on `reqwest`.

Our Rust HTTP code (of which there isn't much) has been using the
Rust `curl` bindings based on the logic that libdnf and libostree
both link to libcurl too.  That's still the case.  However,
since we're already *also* carrying a dependency on a memory-safe
HTTP implementation, let's use it.

Other arguments for this:

- We can drop the code which is removing the vendored copy of the
  libcurl C code from our Rust vendor tree.
- curl actually supports way more than http; for example, it supports
  wacky stuff we *don't* want to support like FTP and `gopher://` etc.
  reqwest is a HTTP-only library and drops all that stuff out of
  our attack surface.
- This code is using the `reqwest::blocking` API, but actually
  since we're using tokio in the future we can change this to be
  more natively async, which should be more performant.
- I want to move more of the memory-unsafe C code out into separate
  isolated subprocesses, keeping the main binary Rust.