libpriv/core: overlay ostree content before running scriptlets
This moves the overlaying of ostree layers earlier in the composing
process, so that their content is present on disk before any
scriptlet is run.
Before this, layered files were added only after `%pre` and `%post`
logic already ran. However, that resulted in troubles when trying
to add configuration fragments which would influence scriptlets behavior
(e.g. `sysusers.d` entries), as the files were not present in the target
envinroment at the time the scripts are executed.
This friction was notably observed with systemd `%post` script, which runs
`systemd-sysusers` and would miss any customization/fragments from ostree
layers.