core: wrap and intercept `groupadd` calls in scriptlets
This re-routes `groupadd` calls in scriptlets, in order to inspect
group creation by RPMs and automatically generate corresponding
sysusers.d fragments.
It uses a new hidden `scriptlet-intercept` subcommand, which can
be re-used in the future to intercept additional commands (e.g.
`useradd`).
For the moment, this logic is experimental and gated behind an
environment feature flag.