install: Add `-y/--assumeyes` option, prompt on tty by default
This is the dual/inverse of
https://github.com/coreos/rpm-ostree/pull/3884/commits/87a1da1fb36a9b23ac9215317ce7580d61aecd7f

A big difference between rpm-ostree and yum/dnf today is
rpm-ostree does *not* prompt when you type `rpm-ostree install foo`;
it just does the operation by default.  One rationale for this is because
it doesn't apply live, it really can't break anything before
the admin has a chance to take action.

Valid cases here include:

 - Discovering that you typo'd the package name
 - Adding the package drags in a lot more dependencies than expected

However, when we added `-A/--apply-live` we broke that guarantee.
Now of course unlike traditional yum we have rollbacks.   But still,
I think in the future we should switch the default to prompting for
confirmation *only* when `-A/--apply-live` is passed.

Add `-y/--assumeyes` matching yum so that users who want that
no prompt behavior can start adding it now.

We avoid breaking backwards compatibility here for *scripts* by
adding `-y` if we detect that `-A` is passed and stdin is not a tty.