Move `container-encapsulate` under `compose`
One really awkward thing right now is that anything on the Rust
side isn't emitted when one runs `rpm-ostree --help`.  Long
term we need to hard port all the CLI to Rust.

In the short term, I wanted to promote/document `container-encapsulate`.
Let's fix the "not in --help" problem by moving it where it probably
should have been to start - under the `compose` verb which is
our "build stuff".