unit: Bump `TimeoutStartSec=5m`
Significantly bump this timeout from the default because
we do a lot of stuff on daemon startup.

Immediate motivation is https://bugzilla.redhat.com/show_bug.cgi?id=2111817

But this is also related to the same problems that motivated
https://github.com/coreos/rpm-ostree/issues/3850
(cc https://github.com/coreos/rpm-ostree/pull/2932 )

We switched from the default DBus timeout of 25 seconds to the
systemd default of 90s; this bumps us all the way up to 5 minutes.

I think the right long term fix is the socket activation, but this
is an easy backportable fix that will hopefully paper over spurious
failures.

(That said, anyone who is hitting this regularly probably has
 a system too slow to really use, but...let's not stand in their
 way)