Stabilize `rpm-ostree apply live`
I think it's time; this has been basically working for a long
time now.  We implicitly stabilized most of this anyways
when we added `rpm-ostree install -A`.

There are a ton more things we can do, but I think
the status quo is useful.  Let's ensure people don't fear to use it.

One specific motivation I have here is to have `yum install foo`
default to `rpm-ostree install -A` and it's all just clearer
if the `apply-live` as a whole is stable.