Add an always-on `container` feature
I want to be able to detect in OCP when we have a new enough
rpm-ostree to use the container path.

I started doing some version-number parsing, but then realized
what we really want is a feature.

(That said, now that I think about it a bit more this feature
 in theory should be dynamic, e.g. the daemon could detect on
 startup whether `/usr/bin/skopeo` exists and is new enough, etc.
 But, no need to block on that right now)