Add `ex deploy-from-self`
This experimental command is intended to be mainly used in OCP
where we're trying to do a cutover to the new format container images.
However, we have the problem of old disk images where the rpm-ostree
there is too old to understand the container bits.  (Not to mention
skopeo is too old)

This code (+ integration test) demonstrates running the new OS
update *as a privileged container* which mounts the host filesystem
and performs the update that way.

A downside of this is that we only write the raw ostree commit;
we don't write the container-ostree refs; in other words, the
next update will need to re-pull the container.  That's...fixable,
it just requires a bit more nontrivial work that we can do later
as an optimization.