container: Add progress spinners to `compose container-encapsulate`
It turns out the part that's really slow right now is that RPM
apparently walks the entire database to implement "find the package
owning /usr/bin/foo".  What we need to do is build up this mapping
as part of the build process, and dump it out to a lookaside
file (e.g. `rpm-ostree compose tree --write-owners-file=foo.db`)
and then have the encapsulation code pick it up.

Alternatively, we could change the encapsulation code to walk
all RPMs, but eh.