core: Get the kernel version from the kernel path
Before commit 80bb67c2c9fb4e4048880963d04d3a4d36c85b9a, rpm-ostree
required there is only one directory in /usr/lib/modules and the
directory represents the installed kernel. However, it was revealed that
3rd-party RPMs can have pre-built modules for several kernel versions,
and the requirement prevented from installing them. The commit was
intended to solve the problem by ignoring directories which do not
contain a file named vmlinuz.

However, the upstream kernel actually produces RPMs without vmlinuz in
/usr/lib/modules, and the logic introduced by the commit prevented them
from being installed.

This change replaces the logic with yet another alternative that
retrieves the kernel version from the kernel file name in the boot
directory. The kernel version can be used to identify the kernel module
directory without requiring that there is only one directory in
/usr/lib/modules or that the directory contains vmlinuz.

An old comment in find_kernel_and_initramfs_in_bootdir says the kernel
installed in Fedora 23 does not contain the version in the path, but
it should not be a problem; looking at the source code, the path
has always contained the version since Fedora 12.