core: Filter for latest when downloading packages
The tightly-scoped function `rpmostree_find_and_download_packages`
function just takes a query and returns fds to downloaded packages. It
does not do any depsolving, so it's up to us to select the exact
package we want to download. This means explicitly telling libdnf we
want the latest version. This matches e.g. `dnf download`.

Closes: #4037