rebuild: Fix logic for container-only handling
Right now `ex rebuild` only works in an (ostree) container and
trying to run it on an ostree host system, despite all the checking
in that function just segfaults.

And the reason here is very illuminating: We were passing `NULL`
for the out parameter for the DBus client.  Let's use that to
instead be the signal that a command is current container *only*.

But, this is really a perfect example of where Rust would have
entirely prevented this problem.