container: Enable wrappers duing transaction
This ensures that we get the same wrappers for `/usr/bin/systemctl`
and `/usr/bin/groupadd` in the container flow that we have
at `compose tree` time.

(Also, it's really confusing that this functionality is separate
 from `cliwrap`...we need to unify these two things)

While I'm here, I hit an exciting bug in the container flow where
I accidentally typo'd a package name, and because we didn't
"undo on drop" we left our wrappers active.  (This doesn't matter
on the compose side as we're always acting on a temporary filesystem)

Bigger picture we should do
https://internals.rust-lang.org/t/should-rust-programs-unwind-on-sigint/13800/11