composepost: Drop unnecessary injection of presets
First motivation: this logic is not idempotent as it should be,
and someone hit this when trying to test out a local commit created
via `compose commit`.

But bigger picture, quite a while ago we did
https://github.com/ostreedev/ostree/commit/73e3ccc401829025a9151fddff29d67a0ac2321d

I think we're at the point where we can hard require that updated
ostree.