treefile: Make `container: true` opt-in to saner defaults
As part of our expansion into a general "dnf image", it now
makes total sense to use what's today `rpm-ostree compose image`
to generate *non*-bootable container images.  For example,
we could use this to generate coreos-assembler.

Then, we have the same compelling features that we use for the
OS that apply there:

- sane change detection
- chunking

We could also use this to generate other containers, such
as our buildroot image.