Always use merge commit for container deployments
In this whole project of "ostree native containers", there's
been some fuzziness as far as whether the container stuff is just
a "transport" or whether it more *replaces* the ostree data.

We definitely have support for the former, which we need.

But there's this confusing leftover logic on the client side
which uses the base ostree commit if we can - which we can *only*
do if there's exactly one container image layer.  Which...is
really a dead approach now that we have "chunked" container images.

Switch to always deploying the merge commit, since that's what
we'll use in all real production scenarios.

This is prep for image layer GC handling - we want to be able to
*always* fetch the container image merge commit.

Currently trying to do that trips up in our integration tests
which use a bare `ostree container encapsulate` and rebase to it.