ci: Run "Build Integration Test Data" GHA privileged
The glib2 shipping in Fedora 37 is hitting the classic seccomp EPERM vs
ENOSYS issue for `close_range` when used via `createrepo_c`.
Interestingly, Fedora 36 carried a patch for this:

https://src.fedoraproject.org/rpms/glib2/c/a2259ad90593383c5ce982fbb233fd3658c0a7a1?branch=f36

But this patch is not carried in Fedora 37, presumably on the basis that
by then hosts should be running a new enough runc to fix

https://github.com/opencontainers/runc/issues/2151

But clearly, that hasn't happened yet for whatever version runc that
moby-engine uses in `ubuntu-latest`.

Hack around this by running the container in privileged mode.