cxxrsutil: Drop use of `&mut` in `gobj_wrap()`
Prep for updating to the newer glib, which seems to change
something related to how Deref works here.

We never needed to cast to a `&mut` reference here - the
semantics of glib-rs all use "interior mutability" (which contrasts
with how cxx.rs tries to propagate C++ `const` with Rust borrows).