core: Allow lockfiles to reference missing package names
I'm trying to reuse the x86_64 lockfile generated during a cosa build to
ensure that multi-arch builds for the same version will match (on
streams where we don't yet have in-git lockfiles). One issue with that
is that some packages are arch-specific (like `bootupd`), so composes on
s390x will fail. This relaxes the lockfile logic so that a locked
package can be ignored if we didn't find that package name at all in the
repos when in non-strict mode.

Lockfile semantics are to restrict the set of packages the depsolver
picks from, not to force the installation of some packages even if
unrequested (this is true even for strict mode). So if a locked package
name doesn't exist at all in the repos, there are no packages to
restrict.

We still warn so that a mistyped package name can more easily be
noticed. We also keep making it a hard error in strict mode since in
that workflow we expect lockfiles to be better managed.