libpriv/kernel: fix kver parsing from vmlinuz in /boot and /usr/lib/ostree-boot
When deriving the kver from the `vmlinuz` file found in `/boot` and
`/usr/lib/ostree-boot`, we need to strip the trailing bootcsum digest
if we added it. This is an issue for Fedora IoT, which still uses
`boot-location: new` (we should probably move to deprecate that option
in favour of just `modules`).

Closes: #4241