compose: Drop support for `source-packages` in lockfiles
We never ended up using this and focus shifted on more automation to
better manage lockfiles. In such cases, it's easier to deal with binary
packages that match exactly what's in the host and what cosa spits out
than source package names, which require some translation and make
override semantics harder to calculate.

Let's just drop support for it to simplify code.

This is mostly a revert of the code that added it, 7979e250 ("lockfile:
Allow locking by source RPM EVR").