core: Further loosen lockfile handling
This is a follow-up to 660c7294 ("core: Allow lockfiles to reference
missing package names"). There, we loosened semantics so that locked
package could be ignored if that package name doesn't exist at all in
the repos.

This patch loosens it further so that we don't immediately error out if
the locked package's NEVRA doesn't exist in the repos and instead rely
on the depsolver to error out for us. We still warn if we couldn't find
a locked package to be helpful.

Note the semantics of the lockfile are still maintained. In relaxed
mode, we exclude all packages of the same name as those found in the
lockfile. In strict mode, we exclude all packages not locked. The key
difference with this patch is that if a locked NEVRA couldn't be found
in the repos, it's only an error if the compose needs that package.

I hit this when trying to reuse an x86_64 lockfile in an s390x rawhide
build. For some reason, `grub2-common.noarch` isn't in the official
repos for s390x, but is in the coreos-pool repo for s390x. I think
the handling of `noarch` packages between the two tools that generate
these repos differs. The end result is that `cosa build` on s390x fails
because it can't find the locked `grub2-common` package, which we don't
even need.

A cleaner approach for what we're trying to do with this stuff (cross-
arch content consistency) is to work with snapshotted repodata. Though
even there, we probably would still want a post-compose check that
versions across arches matched. Another approach would be to teach rpm-
ostree to depsolve for multiple basearches in a single invocation and
output the corresponding lockfiles, but that's more complex.

Anyway, this isn't much work and will enable stronger semantics than
we have right now in RHCOS (where we'll eventually get lockfiles) and
rawhide (where we have no plans to add lockfiles).