treefile: Return `.` instead of `""` for parent directory
This addresses the root cause of
https://github.com/coreos/rpm-ostree/issues/4029

Basically we were calling `set_repos_dir("")` on the libdnf side
and it simply did not handle that, failing to open the directory.

(Though I'm not saying "Closes" since we should still fix the
 file monitoring stuff)

I don't yet know why this seemed to work in some cases but not
others though.