Support `LockLayering=true` config knob
This knob provides an easy way for a sysadmin to disable all package
layering and initramfs customizations without having to muck around
with polkit policies. This has been requested a few times in the past
(and again recently) in environments where we want more safety checks
(e.g. Fedora IoT/RHEL for Edge). There's more we could do of course for
that use case (I think sealing `/etc` has come up, though that probably
belongs better in libostree), but this is low-hanging.

We do this at the upgrader level rather than higher up in the stack to
be sure we catch all paths, but also so that if the feature is enabled
on a system with existing customizations, rpm-ostree still permits doing
a `reset` to bring it back in line.

Closes: #4061