WIP: Temporarily transition rpmdb to bdb
This is aiming to fix https://issues.redhat.com/browse/OCPBUGS-7275

Basically to handle the rpmdb transition from bdb to sqlite seamlessly,
we need to down-convert because for bad historical reasons we
use the *host* rpm stack to do writes; and here the host only
can write to bdb.

So on the next boot after an upgrade with layered packages, we will
be in a situation where the host rpmdb is bdb - this is the same
as a traditional dnf/yum system.

Where things diverge is that there the `rpmdb-migrate.service`
will kick in and mutate the system live.  In our case, such mutation
conflicts with our philosophy and technology.  What will happen
for us instead is that because we always start from the new base
image, we'll detect that host preferred rpmdb is `sqlite`, the
target is `sqlite`, and just proceed with the new base image
sqlite db.

I got myself confused a bit into thinking our bdb database would
persist, but it doesn't.