daemon: Use a socket in `/run`, require non-abstract
This fixes https://bugzilla.redhat.com/show_bug.cgi?id=2169622

The new glib changed to use non-abstract sockets by default,
which broke us because we'd slowly grown more isolation, specifically
the daemon has been using `PrivateTmp` for a while and we'd
been relying on abstract sockets to pierce that.

Change to use `/run` which should always be shared by client
and daemon.  While we're here, make it a well-known static path
because there can be only one transaction at a time.

We also do need to explicitly make the socket world-accessible
because that was the semantics of the previous abstract socket.

Also, plug leaks here by ensuring we call `g_dbus_server_stop()`.

(But really we should also change this to be crash safe, probably
 with a temporary directory, but that's a larger change)