build: Allow GLib 2.70, also `-Wno-error=deprecated-declarations`
First bump the max glib to 2.70 since otherwise we get a spam
of warnings from the new libglnx which feature-checks glib and
invokes `g_steal_fd()` if available.

(TODO: change libglnx to suppress glib max-available warnings
 when `-DGLIB_VERSION_MAX_ALLOWED` is present but we've feature
 checked)

Also while we're here, we don't want to warn on deprecations even
in CI, because it can change outside of our world.