Make output handling thread-local
This is needed because for transactions (which are always run in
a thread today) we want output to go to the transaction's DBus
progress.  But for other methods which are not transactions, we don't
have a channel for status reporting, so output needs to
continue to go to the journal.   If we mix these two things due to
concurrent method invocations, the client may get confused and crash.

Closes: https://github.com/coreos/rpm-ostree/issues/4284