Add an `exp` verb
Often now when I want to add some quick new experimental entrypoint
the ergonomic hit from having to write C++ and pass through the
arguments via the existing `ex` is painful.

Actually today, almost all the `ex` bits are in Rust natively,
so we could fix that.  However, that'd be some nontrivial code
churn.

Add a new `exp` verb that is natively in Rust to start, and
add a `container diff` command that just exposes the logic
we landed in ostree-ext for this.  I will use this to
evaluate the new chunking work.