core: Error out instead of aborting on reinstalls
I think originally something higher level was intending to
catch this "reinstall exact base package version" case, but
it's way simpler to just return a clear error here instead
of crashing.

Assertions should only be used for things we're *pretty sure*
can't be reached and particularly when it may be a
hard to recover from internal consistency error.  That's
not the case here.

Closes: https://github.com/coreos/rpm-ostree/issues/4141