Add `kola` tests and fix `Name & Summary` search
`kola` tests added to test basic search functionality (name, summary, name &
summary, and glob matches).

Small fix for the duplicate removal feature. The `set` used to store
existing query results needed to be passed as a pointer to the
`query_results_to_builder` function. This allowed the data within the
`set` to persist outside the function call.

A large was change made to the `Name & Summary` search feature. Previously,
the function would only return results that contained ALL search terms
within BOTH the name and summary. With this commit, results in which
search terms match EITHER the name or summary will be returned.
This better matches `dnf`/`yum` CLI functionality.