compose-image: Add `--initialize-mode`
A long time ago I did
https://github.com/containers/skopeo/pull/1757/commits/08b27fc50e4da2501124d573b4a449dfeee260ad
in preparation for this change.

Basically this is what we really want to be the default, but
couldn't at the time: "Create a new image if one doesn't exist".

For completeness though, we also add support for `always`
(which is the existing `--initialize`) as well as `--never`
(which ensures we never overwrite an existing image, in case
 someone cares).

However in testing this out: it basically works OK with
a registry transport which is the big one, but:

- It just plain doesn't work with `.ociarchive` due to skopeo bugs
- It even more unfortunately doesn't work with `oci` directories
  and a target image reference;
  e.g. `--format=oci manifest.yaml oci:foo:sometag`