libpriv/core: Tolerate rpmdb vs on-disk file type mismatch
Trying to `override replace container-selinux` breaks with:

```
error: unlinkat(usr/etc/selinux/targeted/active/modules/200/container): Is a directory
```

The reason is that this path is declared as a `%ghost` file in the
package, but is in fact a directory. We could fix the packaging there,
but in fact librpm tolerates these mismatches simply by going off of the
file type of the target path on-disk. So let's match that.