Add `kola` test for local dir container deployment
This change adds a test for deploying FCOS from a container image stored
in a local directory. The test is needed for a change in ostree-rs-ext
which adds support for the `dir` transport.

In this test, the encapsulate function is used to store the current state
of the system to a local directory and is deployed to a temporary sysroot.