libpriv/kernel: Add padding between dracut initramfs and random CPIO
Any arbitrary amount of padding between concatenated CPIOs is allowed:

https://www.kernel.org/doc/Documentation/early-userspace/buffer-format.txt

However, some initramfs decompressors in the kernel expect a minimal
amount of padding to detect EOF and allow the kernel outer loop to try
out other decompressors for the next payload. This is the case for lz4.
This upstream patch allows the lz4 decompressor to treat padding as EOF:

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=2c484419efc09e7234c667aa72698cb79ba8d8ed

But it assumes that the padding is at least 4 bytes.

Since padding doesn't hurt regardless of the compression used, let's just
unconditionally add it.

Fixes https://github.com/coreos/rpm-ostree/issues/4683.

Co-authored-by: Colin Walters <walters@verbum.org>